backend/requirements.txt
Django==4.2.1
django-extensions==3.2.1
djangorestframework==3.14.0
django-rest-passwordreset==1.3.0
django-cors-headers==3.14.0
python-decouple==3.8
psycopg2-binary==2.9.6